+++ {{ cookiecutter.project_name }}/composition_repository/lotus/__init__.py
@@ -1,0 +1,6 @@
+DEPENDENCIES = [
+    "ckeditor",
+    "dal",
+    "taggit",
+    "smart_media",
+]