+++ {{ cookiecutter.project_name }}/composition_repository/lotus/settings.py
@@ -1,0 +1,18 @@
+from project_composer.marker import EnabledApplicationMarker
+
+from lotus.contrib.django_configuration import LotusDefaultSettings
+
+
+class LotusSettings(LotusDefaultSettings, EnabledApplicationMarker):
+    """
+    Lotus settings
+    """
+
+    @classmethod
+    def setup(cls):
+        super(LotusSettings, cls).setup()
+
+        cls.INSTALLED_APPS.extend([
+            "view_breadcrumbs",
+            "lotus",
+        ])