+++ {{ cookiecutter.project_name }}/composition_repository/lotus/urls.py
@@ -1,12 +1,11 @@
 from django.conf.urls.i18n import i18n_patterns
 from django.urls import include, path
-
 from project_composer.marker import EnabledApplicationMarker
 
 
-class CmsBaseUrls(EnabledApplicationMarker):
+class LotusUrls(EnabledApplicationMarker):
     """
-    CMS urls
+    Lotus urls
     """
 
     def load_urlpatterns(self, urlpatterns):
@@ -15,11 +14,6 @@ class CmsBaseUrls(EnabledApplicationMarker):
         """
         urlpatterns = super().load_urlpatterns(urlpatterns)
 
-        if self.settings.ENABLE_I18N_URLS:
-            return i18n_patterns(
-                path("", include("cms.urls")),
-            ) + urlpatterns
-        else:
-            return [
-                path("", include("cms.urls")),
-            ] + urlpatterns
+        return i18n_patterns(
+            path("blog/", include("lotus.urls")),
+        ) + urlpatterns