+++ {{ cookiecutter.project_name }}/composition_repository/taggit/settings.py
@@ -1,0 +1,15 @@
+from project_composer.marker import EnabledApplicationMarker
+
+
+class TaggitSettings(EnabledApplicationMarker):
+    """
+    django-taggit settings
+    """
+
+    @classmethod
+    def setup(cls):
+        super(TaggitSettings, cls).setup()
+
+        cls.INSTALLED_APPS.extend([
+            "taggit",
+        ])